backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.20
langgraph==0.0.25
langchain-openai==0.0.8
python-dotenv==1.0.0
pydantic==2.6.0
supabase==2.3.4
jupyter==1.0.0
